displays card from state store using pre-fetch from endpoint
