+++ client/src/components/atom/btn.jsx
@@ -1,7 +1,9 @@
 import { Button } from "react-bootstrap";
 
+import styles from "../../styles/components/atom/btn.module.css";
+
 export const btn = (variant, title) => (
-	<Button type="submit" variant={variant}>
+	<Button className={styles.btn} type="submit" variant={variant}>
 		{title}
 	</Button>
 );
