+++ client/src/components/organism/Cards.jsx
@@ -1,20 +1,38 @@
-import Article from "./Article.jsx";
+import { Card } from "react-bootstrap";
+
+import { btn, spinner } from "../atom";
 
 import useInfiniteScroll from "../../hooks/useInfiniteScroll.jsx";
 
+import styles from "../../styles/components/organism/Cards.module.css";
+
 const Cards = () => {
-	// NOTE: this will load the data from the endpoint
-
-	// TODO: the Article component can now be deleted and removed
-	// TODO: all styling can be done from this component
-	// TODO: all the styling can be drawen from the CardMolecule
-
-	// TODO: needing to see what molecules and atoms I need for this component
+	const data = useInfiniteScroll("/card");
 
 	return (
-		<div>
-			<Article />
-		</div>
+		<>
+			{data.length === 0 ? (
+				spinner()
+			) : (
+				<div className={styles.container}>
+					{data.map(({ id, title, description, image }) => (
+						<div key={id}>
+							<Card className={styles.card}>
+								<Card.Title className={styles.title}>{title}</Card.Title>
+								<Card.Img className={styles.image} variant="top" src={image} />
+								<Card.Text className={styles.description}>
+									{description}
+								</Card.Text>
+								<div className={styles.description}>
+									{btn("danger", "Delete")}
+									{btn("primary", "Update")}
+								</div>
+							</Card>
+						</div>
+					))}
+				</div>
+			)}
+		</>
 	);
 };
