+++ client/src/styles/components/atom/btn.module.css
@@ -1,0 +1,3 @@
+.btn {
+	margin: 5px;
+}
